Addressed issue #1. Now the page asks the user to confirm for big files and refuses to compare extra-large files. Moreover, the algorithm now runs asynchronously so that there's less chance to block the user's browser / system.
Implemented a new automation feature: producer of automation tools can use COM or whatever interop technology of browsers, as long as it supports executing JavaScript code, can call `setFileNames` and then `doCompare` to provide the file names and content to be compared. The page should display as if the user has selected the corresponding files and tapped "compare". Note: this automation feature does not check validity. The caller is responsible for verifying the parameters are correct and the content to be compared is not too large.